+++ front/src/app/modules/compute-type/compute-type-routing.module.ts
@@ -16,6 +16,7 @@ const routes: Routes = [
     component: ComputeTypeListComponent,
     resolve: {
       computeTypes: ComputeTypesResolver,
+      resources: OvhResourceResolver,
     }
   },
   {