+++ front/src/app/modules/sessions-resources/interfaces/session-with-resources.ts
@@ -10,3 +10,8 @@ export interface ISessionWithResources extends IAdminSession {
   course: IAdminCourse;
   resources: IResourceDescription[];
 }
+
+export interface IResourceRegions {
+  ovhResourceId: number;
+  regions: string[];
+}